% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cw.R
\name{create_cw}
\alias{create_cw}
\title{Create a Crosswalk between Two Sets of Spatial Features (shapefiles)}
\usage{
create_cw(
  from_sf,
  to_sf,
  wts_sf,
  wt_var_name,
  check_that_wts_cover_from_and_to = TRUE,
  wts_check_buffer_frac = 0.001,
  check_for_ak = TRUE,
  check_for_hi = TRUE
)
}
\arguments{
\item{from_sf}{An sf object representing the source features. Must
have columns named "geoid" and "geometry" with geometry of class
'sfc_MULTIPOLYGON'.}

\item{to_sf}{An sf object representing the target features. Must
have columns named "geoid" and "geometry" with geometry of class
'sfc_MULTIPOLYGON'.}

\item{wts_sf}{An sf object representing the weight features. Must
have columns named "geoid" and "geometry" with geometry of class
'sfc_MULTIPOLYGON'. It must have a column used for weighting. The
name for this weighting column must be passed to
\code{wt_var_name}. For area-based weighting, use a constant (e.g.,
\code{1}) as the weighting variable for all polygons in \code{wts_sf}.}

\item{wt_var_name}{A character string specifying the name of the
column in \code{wts_sf} to be used for weighting (required).}

\item{check_that_wts_cover_from_and_to}{A logical value indicating
whether to check that \code{wts_sf} completely covers both \code{from_sf}
and \code{to_sf} (default is TRUE).}

\item{wts_check_buffer_frac}{A numeric value indicating how much to
buffer the weights shapefile when checking that it covers the
from and to shapefiles. Default is 0.001, or 0.1\%}

\item{check_for_ak}{A logical value indicating whether to check for
and handle Alaska specifically (default is TRUE).}

\item{check_for_hi}{A logical value indicating whether to check for
and handle Hawaii specifically (default is TRUE).}
}
\value{
A data.table with the crosswalk between \code{from_sf} and \code{to_sf}.
}
\description{
This function generates a crosswalk between two sets of spatial features
(\code{from_sf} and \code{to_sf}) using a third set of spatial features (\code{wts_sf})
for weighting. The crosswalk maps \code{from_sf} to \code{to_sf} based on the
intersection with \code{wts_sf}, providing weights for how much of \code{from_sf}
intersects with \code{to_sf} relative to \code{wts_sf}.
}
\details{
The function performs the following steps:
\enumerate{
\item \strong{Input Validation:} Checks the classes and column names of the input sf
objects.
\item \strong{Data Preparation:} Converts sf objects to data.tables and renames
columns for consistency.
\item \strong{Alaska and Hawaii Handling:} If \code{check_for_ak} or \code{check_for_hi} is
TRUE, the function identifies and transforms the coordinate reference
systems (CRS) of Alaska and Hawaii polygons for accurate calculations.
\item \strong{Crosswalk Computation: (via \code{create_cw_worker()})}:
\itemize{
\item Identifies 1:1 mappings where \code{to_sf} completely covers \code{from_sf}.
\item For the remaining \code{from_sf} polygons, it finds the \code{wts_sf} polygons
covered by both \code{from_sf} and \code{to_sf}.
\item Calculates the intersection of the relevant polygons from \code{from_sf},
\code{to_sf}, and \code{wts_sf}.
\item Determines the share of the \code{wt_var} for each \code{wts_sf} polygon that
relates to the intersection with \code{from_sf} and \code{to_sf}.
}
\item \strong{Output:} Returns a data.table containing the crosswalk with the
following columns:
\itemize{
\item \code{from_geoid}: The "geoid" from \code{from_sf}.
\item \code{to_geoid}: The "geoid" from \code{to_sf}.
\item \code{wt_var_name}: The weight variable from \code{wts_sf} (renamed from
\code{wt_var_name}).
\item \code{afact}: The allocation factor from \code{from_geoid} to \code{to_geoid} that
represents the share of \code{from_geoid} allocated to \code{to_geoid}. Note
that the sum of \code{afact} for each \code{from_geoid} is 1, meaning that 100\%
of each \code{from_geoid} is allocated to a \code{to_geoid}.
}
}
}
\examples{
\dontrun{
library(geolinkr)
library(sf)
library(tigris)

ct_cnty20 <- tigris::counties(state = "CT", year = 2020) |>
  sf::st_transform(crs = 5070) |>
 _[, c("GEOID", "geometry")]
names(ct_cnty20) <- c("geoid", "geometry")

# 2023 CT counties -- target shapefile
ct_cnty23 <- tigris::counties(state = "CT", year = 2023) |>
  sf::st_transform(crs = 5070) |>
 _[, c("GEOID", "geometry")]
names(ct_cnty23) <- c("geoid", "geometry")

# 2020 CT census tracts -- wts shapefile (with hh2020 as the weight variable)
ct_tracts20 <- readRDS(url("https://github.com/ChandlerLutz/ct-shps/blob/main/ct_tracts20_sf.rds?raw=1")) |>
 sf::st_transform(crs = 5070) |>
 _[, c("GEOID", "hh2020", "geometry")]
names(ct_tracts20) <- c("geoid", "hh2020", "geometry")

cw_ct_cnty20_cnty23 <- create_cw(
 from_sf = ct_cnty20,
 to_sf = ct_cnty23,
 wts_sf = ct_tracts20,
 wt_var_name = "hh2020",
 check_that_wts_cover_from_and_to = FALSE
)

}

}
