% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cw_worker.R
\name{create_cw_worker}
\alias{create_cw_worker}
\title{Create Crosswalk (Worker Function)}
\usage{
create_cw_worker(dt_from, dt_to, dt_wts)
}
\arguments{
\item{dt_from}{A data.table containing the "from" polygons with columns:
\itemize{
\item \code{from_geoid}: Character vector representing the unique IDs of the "from" polygons.
\item \code{geometry}: An sf geometry column containing the "from" polygons.
}}

\item{dt_to}{A data.table containing the "to" polygons with columns:
\itemize{
\item \code{to_geoid}: Character vector representing the unique IDs of the "to" polygons.
\item \code{geometry}: An sf geometry column containing the "to" polygons.
}}

\item{dt_wts}{A data.table containing the weight polygons with columns:
\itemize{
\item \code{wts_geoid}: Character vector representing the unique IDs of the weight polygons.
\item \code{wt_var}: Numeric vector representing the weight variable associated with each weight polygon.
\item \code{geometry}: An sf geometry column containing the weight polygons.
}}
}
\value{
A data.table with the following columns representing the crosswalk:
\itemize{
\item \code{from_geoid}: Character vector of "from" polygon IDs.
\item \code{to_geoid}: Character vector of "to" polygon IDs.
\item \code{wt_var}: Numeric vector representing the weight variable.
\item \code{afact}: Numeric vector representing the area factor (normalized weights).
}
}
\description{
This function is a worker function used to create a crosswalk between two sets of
polygons (\code{dt_from} and \code{dt_to}) using a third set of polygons (\code{dt_wts}) for
weighting.
}
\details{
This function performs the following steps:
\enumerate{
\item \strong{Input Validation:} Checks if the inputs are data.tables and contain the required columns.
\item \strong{CRS Check:} Ensures that all input data.tables have the same Coordinate Reference System (CRS).
\item \strong{Empty Geometry Handling:}  Removes any empty geometries from the input data.tables.
\item \strong{1:1 Mapping:} Identifies cases where \code{dt_to} polygons completely cover \code{dt_from} polygons and creates a 1:1 mapping.
\item \strong{Weighted Crosswalk:} For the remaining polygons, calculates the intersection between \code{dt_from}, \code{dt_to}, and \code{dt_wts}
and determines the weights based on the area of intersection relative to the weight polygons.
}
}
