% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_visual_cw_2020_2022}
\alias{ct_visual_cw_2020_2022}
\title{Connecticut County Visual Crosswalk (2020 - 2022)}
\format{
A data frame with 18 rows and 2 variables:
\itemize{
\item \code{cntyfp2020}: County FIPS code for 2020.
\item \code{cntyfp2022}: County FIPS code for 2022
}
}
\source{
Chandler Lutz. This dataset is for illustrative purposes only and does not represent actual changes in Connecticut county FIPS codes.
}
\usage{
ct_visual_cw_2020_2022
}
\description{
A dataset providing a visual crosswalk for Connecticut counties between 2020 and 2022.
}
\examples{
data(ct_visual_cw_2020_2022)
head(ct_visual_cw_2020_2022)
}
\keyword{datasets}
