% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_nonoverlapping_square_polygons}
\alias{gen_nonoverlapping_square_polygons}
\title{Generate Non-Overlapping Square Polygons}
\usage{
gen_nonoverlapping_square_polygons(
  num_squares,
  side_length,
  squares_per_row,
  crs = NULL,
  start_at_xy = c(0, 0)
)
}
\arguments{
\item{num_squares}{The total number of squares to generate.}

\item{side_length}{The length of the sides of each square.}

\item{squares_per_row}{The maximum number of squares to place in a
horizontal row before starting a new row.}

\item{crs}{An integer representing the coordinate reference system
(CRS) for the generated polygons.  If NULL (default), no CRS is
set.}

\item{start_at_xy}{A numeric vector of length 2 specifying the x
and y coordinates of the bottom-left corner of the starting
square. Default is c(0, 0).}
}
\value{
An sf object containing the generated non-overlapping
square polygons as MULTIPOLYGONs.
}
\description{
This function generates a set of non-overlapping square polygons arranged in a grid-like pattern.
The squares are stacked in rows, with the option to specify the number of squares per row and the
starting position of the grid.
}
\details{
The function calculates the position of each square based
on the \code{side_length}, \code{squares_per_row}, and \code{start_at_xy}
parameters. The \code{crs} argument allows you to set a coordinate
reference system for the generated polygons.
}
\examples{
# Generate 12 squares with a side length of 2, arranged in rows of 4,
# starting at (0, 0)
squares_1 <- gen_nonoverlapping_square_polygons(
  num_squares = 12, 
  side_length = 2, 
  squares_per_row = 4
)

# Generate 8 squares with a side length of 1, arranged in rows of 3, 
# starting at (5, 10), and using WGS 84 (EPSG:4326)
squares_2 <- gen_nonoverlapping_square_polygons(
  num_squares = 8, 
  side_length = 1, 
  squares_per_row = 3, 
  crs = 4326, 
  start_at_xy = c(5, 10)
)
}
