% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{polygon_intersects_with_us_state}
\alias{polygon_intersects_with_us_state}
\title{Check Polygon Intersections with US States}
\usage{
polygon_intersects_with_us_state(geometry)
}
\arguments{
\item{geometry}{An \code{sf} object with geometries of class
\code{sfc_MULTIPOLYGON}. The object must have a column named
'geometry'.}
}
\value{
An integer vector, where 1 indicates that the corresponding
polygon intersects with a US state, and 0 indicates no
intersection.
}
\description{
This function checks if the geometries in a given \code{sf} object intersect with US states.
}
\details{
The function first verifies that the input has the correct
geometry type (\code{sfc_MULTIPOLYGON}). It then transforms the input
geometry to CRS 5070 for accurate spatial operations, converts
the geometry to a \code{data.table}, and checks for intersections with
US state geometries.
}
\examples{
stopifnot(
 polygon_intersects_with_us_state(geolinkr::us_states_sf$geometry) == rep(1L, 51)
)

stopifnot(
 polygon_intersects_with_us_state(geolinkr::puerto_rico_sf$geometry) == 0L
)
}
