% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{polygon_is_ak}
\alias{polygon_is_ak}
\title{Identify Alaska Polygons}
\usage{
polygon_is_ak(geometry)
}
\arguments{
\item{geometry}{An \code{sf} object with geometries of class \code{sfc_MULTIPOLYGON}. The object must have a column named 'geometry'.}
}
\value{
An integer vector, where 1 indicates that the corresponding polygon intersects with Alaska, and 0 indicates no intersection.
}
\description{
This function checks if the geometries in a given \code{sf} object intersect with the state of Alaska.
}
\examples{
\dontrun{
# Identify Alaska polygons in US states
library(data.table); library(sf)
us_states <- geolinkr::us_states_sf \%>\%
  as.data.table() \%>\%
  .[, is_ak_st_code := fifelse(stusps == "AK", 1L, 0L)] \%>\%
  .[, is_ak := polygon_is_ak(geometry)]
stopifnot(us_states[, all(is_ak_st_code == is_ak)])
}
}
