% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{puerto_rico_sf}
\alias{puerto_rico_sf}
\title{Puerto Rico Geometry}
\format{
A simple feature collection (\code{sf} object) with 1 row and 2 columns:
\itemize{
\item \code{statefp}: A character vector representing the FIPS code for Puerto Rico.
\item \code{stusps}: A character vector with the USPS abbreviation for Puerto Rico.
\item \code{geometry}: A \code{MULTIPOLYGON} geometry column representing the boundary of Puerto Rico.
}
}
\source{
US Census Bureau.
}
\usage{
puerto_rico_sf
}
\description{
A spatial dataset containing the boundary of Puerto Rico.
}
\details{
The dataset is projected in \strong{NAD83 / Conus Albers} (EPSG:5070) for consistent spatial analysis with other US territories. It includes a single feature with a \code{MULTIPOLYGON} geometry representing the boundary of Puerto Rico.
}
\keyword{datasets}
