% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_states_sf}
\alias{us_states_sf}
\title{US States Geometries}
\format{
A simple feature collection (\code{sf} object) with 51 rows and 2 columns:
\itemize{
\item \code{statefp}: A character vector representing the FIPS code for each state.
\item \code{stusps}: A character vector with the two-letter USPS abbreviation for each state.
\item \code{geometry}: A \code{MULTIPOLYGON} geometry column representing the boundaries of each state.
}
}
\source{
US Census Bureau.
}
\usage{
us_states_sf
}
\description{
A spatial dataset containing the boundaries of the 51 US states (including the District of Columbia).
}
\details{
The dataset is projected in \strong{NAD83 / Conus Albers} (EPSG:5070) for accurate spatial analysis across the contiguous US. It includes 51 features (50 states plus the District of Columbia) with geometries of class \code{MULTIPOLYGON}.
}
\keyword{datasets}
